fastapi-server/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
httpx  # optional if you're calling Qdrant via HTTP
